+++ lib/views/components/footer.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class Footer extends StatelessWidget {
+  const Footer({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
